requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
environs==14.1.1
Flask==3.1.0
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==4.0.0
multidict==6.4.3
pillow==11.2.1
priority==2.0.0
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
Quart==0.20.0
quart-cors==0.8.0
requests==2.32.3
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
